Module frc46_token::token::types

Structs

BurnFromParams

Instruction to burn an amount of tokens from another address

BurnFromReturn

The updated value after a delegated burn

BurnParams

Instruction to burn an amount of tokens

BurnReturn

The updated value after burning

DecreaseAllowanceParams

Instruction to decrease an allowance between two addresses

GetAllowanceParams

Params to get allowance between to addresses

IncreaseAllowanceParams

Instruction to increase an allowance between two addresses

MintIntermediate

Intermediate data used by mint_return to construct the return data

MintReturn

Return value after a successful mint. The mint method is not standardised, so this is merely a useful library-level type, and recommendation for token implementations.

RevokeAllowanceParams

Instruction to revoke (set to 0) an allowance

TransferFromIntermediate

Intermediate data used by transfer_from_return to construct the return data

TransferFromParams

Instruction to transfer tokens between two addresses as an operator

TransferFromReturn

Return value after a successful delegated transfer

TransferIntermediate

Intermediate data used by transfer_return to construct the return data

TransferParams

Instruction to transfer tokens to another address

TransferReturn

Return value after a successful transfer

Traits

FRC46Token

A standard fungible token interface allowing for on-chain transactions that implements the FRC-0046 standard. This represents the external interface exposed to other on-chain actors

Type Aliases

AllowanceReturn

BalanceReturn

DecreaseAllowanceReturn

GranularityReturn

IncreaseAllowanceReturn

RevokeAllowanceReturn

TotalSupplyReturn